requirements.txt
fastapi[standard]
pydantic
httpx
uvicorn[standard]
python-dotenv
beautifulsoup4==4.12.2